requirements.txt
SQLAlchemy
mysql-connector-python
